👞👦 Checkpoint ./commands/general/google-autofill.js:15834686/200 ./commands/general/cmdsuggest.js:15834686/1440

+++ commands/general/google-autofill.js
@@ -7,6 +7,8 @@ module.exports = {
     category: 'general',
     usage: '<query>',
     run: async (client, message, args) => {
+      
+      if(!message.channel.nsfw) message.reply('please use this command in a NSFW channel!') 
 
         message.delete()
